rename: LoginSlice to LoginPageSlice
